Create Crawler1.py
This is what I have so far that works relatively well. The only thing that I notice that needs fixing now is that bed bath always returns a null value. I want for it to return 1 or 2 values for bed and or bath.

It is also still in it's test shell because I don't want to commit to this for a week. Right now it's set to get 4 craigslist entries.